backend/services/auth_service/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
bcrypt
pydantic
email-validator
PyJWT
